Restructure actions folder
