+++ src/actions/actionTypes.js
@@ -1,0 +1,3 @@
+export const SIGNUP_LOAD = "SIGNUP_LOAD";
+export const SIGNUP_SUCCESS = "SIGNUP_SUCCESS";
+export const SIGNUP_FAILURE = "SIGNUP_FAILURE";
